Use the container name in the exec call
We were getting wonderful 400 errors from  our calls to exec when there
was another container in the debug pod. Why? Because the API didn't
know which container to attach to. Of course, it could have told us
this...

We want to exec into the debug container, so use that container's name
(which is the same as the pod name) when doing the exec.